+++ tests/deprecated_api/test_remove_1-5.py
@@ -32,9 +32,18 @@ from tests.helpers.utils import no_warning_call
 
 def test_v1_5_0_model_checkpoint_save_checkpoint():
     model_ckpt = ModelCheckpoint()
-    model_ckpt.save_function = lambda *_, **__: None
+    trainer = Trainer()
+    trainer.save_checkpoint = lambda *_, **__: None
     with pytest.deprecated_call(match="ModelCheckpoint.save_checkpoint` signature has changed"):
-        model_ckpt.save_checkpoint(Trainer(), object())
+        model_ckpt.save_checkpoint(trainer, object())
+
+
+def test_v1_5_0_model_checkpoint_save_function():
+    model_ckpt = ModelCheckpoint()
+    with pytest.deprecated_call(match="Property `save_function` in `ModelCheckpoint` is deprecated in v1.3"):
+        model_ckpt.save_function = lambda *_, **__: None
+    with pytest.deprecated_call(match="Property `save_function` in `ModelCheckpoint` is deprecated in v1.3"):
+        _ = model_ckpt.save_function
 
 
 @mock.patch('pytorch_lightning.loggers.wandb.wandb')